[Update] todo와 category 도메인 분리
