+++ src/main/java/com/devita/domain/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.devita.domain.todo.repository;
+package com.devita.domain.category.repository;
 
-import com.devita.domain.todo.domain.Category;
+import com.devita.domain.category.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
